+++ src/game/GameServices.h
@@ -12,7 +12,7 @@ struct GameDataDb;
 
 namespace LogBus { struct Logger; }
 class ScriptEventBus;
-namespace engine { class IAssetStore; }
+namespace engine { class IAssetStore; class IRandom; class ITimeSource; }
 
 struct GameServices {
     const GameConfigData& config;
@@ -20,15 +20,21 @@ struct GameServices {
     LogBus::Logger& log;
     ScriptEventBus& events;
     engine::IAssetStore& assets;
+    engine::IRandom& rng;
+    engine::ITimeSource& time;
 
     GameServices(const GameConfigData& cfg,
                  GameDataDb& db,
                  LogBus::Logger& logger,
                  ScriptEventBus& eventBus,
-                 engine::IAssetStore& assetStore)
+                 engine::IAssetStore& assetStore,
+                 engine::IRandom& random,
+                 engine::ITimeSource& timeSource)
         : config(cfg)
         , dataDb(db)
         , log(logger)
         , events(eventBus)
-        , assets(assetStore) {}
+        , assets(assetStore)
+        , rng(random)
+        , time(timeSource) {}
 };